Correcting def_nei_dist as double values

+++ include/Grid2d.hpp
@@ -120,7 +120,7 @@ public:
 
 protected:
 	static const int def_nei[8][2];
-	static const int def_nei_dist[8];
+	static const double def_nei_dist[8];
 	int _grid_width;            /**< the number of cells on the width of the grid*/
 	int _grid_height;           /**< the number of cells on the height of the grid*/
 	Eigen::Vector2d _cell_size; /**< the size of a cell on the plane*/
